+++ app/javascript/stylesheets/layout/header.scss
@@ -106,3 +106,15 @@
   height: 75vh;
   display: block;
 }
+
+.pagetop-btn {
+  @apply fixed bottom-14 right-14 border-2 border-solid border-yellow-500 rounded-full h-11 w-11;
+
+  i {
+    @apply text-yellow-500 text-2xl;
+  }
+}
+
+.pagetop-btn {
+  transition: opacity 1s ease-in-out;
+}
