+++ app/javascript/view-event/index.js
@@ -1,2 +1,2 @@
 import "./task_all_btn";
-import "./hamburger_btn";
+import "./layout_index_event";
