+++ app/views/shared/_header.html.erb
@@ -72,5 +72,7 @@
       </li>
     </ul>
   </div>
+  <button class="pagetop-btn">
+    <i class="fas fa-chevron-up"></i>
+  </button>
 </header>
-
